+++ Plutarch/BitString.hs
@@ -13,10 +13,10 @@ module Plutarch.BitString (
 ) where
 
 import GHC.Generics (Generic)
-import Plutarch.Builtin (PBuiltinList)
 import Plutarch.Builtin.Bool (PBool, pif)
+import Plutarch.Builtin.ByteString (PByteString)
+import Plutarch.Builtin.Data (PBuiltinList)
 import Plutarch.Builtin.Integer (PInteger)
-import Plutarch.ByteString (PByteString)
 import Plutarch.Internal.Eq (PEq)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
 import Plutarch.Internal.Ord (POrd ((#<)))
