--- Plutarch/Builtin.hs-boot
@@ -1,8 +1,0 @@
-module Plutarch.Builtin (PData) where
-
-import Data.Kind (Type)
-import Plutarch.Internal.Term (S)
-
-type role PData nominal
-type PData :: S -> Type
-data PData s
