+++ Plutarch/Crypto.hs
@@ -15,7 +15,7 @@ module Plutarch.Crypto (
 ) where
 
 import Plutarch.Builtin.Bool (PBool)
-import Plutarch.ByteString (PByteString)
+import Plutarch.Builtin.ByteString (PByteString)
 import Plutarch.Internal.Term (Term, (:-->))
 import Plutarch.Unsafe (punsafeBuiltin)
 import PlutusCore qualified as PLC
