+++ Plutarch/DataRepr/Internal.hs
@@ -71,6 +71,7 @@ import Plutarch.Builtin (
  )
 import Plutarch.Builtin.Bool (PBool, pif)
 import Plutarch.Builtin.Integer (PInteger)
+import Plutarch.Builtin.Unit (PUnit (PUnit))
 import Plutarch.DataRepr.Internal.HList (
   HRec (HCons, HNil),
   HRecGeneric (HRecGeneric),
@@ -130,7 +131,6 @@ import Plutarch.TryFrom (
   ptryFrom',
   pupcast,
  )
-import Plutarch.Unit (PUnit (PUnit))
 import Plutarch.Unsafe (punsafeCoerce)
 
 {- | A "record" of `exists a. PAsData a`. The underlying representation is
