+++ Plutarch/FFI.hs
@@ -31,10 +31,11 @@ import Generics.SOP.Type.Metadata (
   ConstructorName,
   DatatypeInfo (ADT, Newtype),
  )
-import Plutarch.Builtin (PData)
 import Plutarch.Builtin.Bool (PBool)
+import Plutarch.Builtin.ByteString (PByteString)
+import Plutarch.Builtin.Data (PData)
 import Plutarch.Builtin.Integer (PInteger)
-import Plutarch.ByteString (PByteString)
+import Plutarch.Builtin.Unit (PUnit)
 import Plutarch.Internal.Eq (PEq, (#==))
 import Plutarch.Internal.Generic (PCode)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
@@ -71,7 +72,6 @@ import Plutarch.Maybe (PMaybe (PJust, PNothing))
 import Plutarch.Script (Script (Script))
 import Plutarch.Show (PShow)
 import Plutarch.String (PString)
-import Plutarch.Unit (PUnit)
 import PlutusTx.Builtins.Internal (BuiltinBool, BuiltinByteString, BuiltinData, BuiltinUnit)
 import PlutusTx.Code (CompiledCode, CompiledCodeIn (DeserializedCode), getPlc)
 import PlutusTx.Prelude (BuiltinString)
