+++ Plutarch/Internal/Eq.hs
@@ -1,7 +1,17 @@
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE UndecidableInstances #-}
+
 module Plutarch.Internal.Eq (
   PEq (..),
 ) where
 
+import Plutarch.Builtin.Bool
+import Plutarch.Builtin.ByteString
+import Plutarch.Builtin.Data
+import Plutarch.Builtin.Integer (PInteger)
+import Plutarch.Builtin.Unit
+
+import Data.Kind (Type)
 import Data.List.NonEmpty (nonEmpty)
 import Generics.SOP (
   All,
@@ -14,14 +24,11 @@ import Generics.SOP (
   ccompare_NS,
   hcliftA2,
  )
-import Plutarch.Builtin.Bool (
-  PBool (PFalse, PTrue),
-  pif',
-  pnot,
-  (#&&),
- )
-import Plutarch.Builtin.Integer (PInteger)
 import Plutarch.Internal.Generic (PCode, PGeneric, gpfrom)
+import {-# SOURCE #-} Plutarch.Internal.IsData
+import Plutarch.Internal.Lift
+import Plutarch.Internal.ListLike
+import Plutarch.Internal.Other
 import Plutarch.Internal.PLam (plam)
 import Plutarch.Internal.PlutusType (
   PlutusType,
@@ -29,6 +36,7 @@ import Plutarch.Internal.PlutusType (
   pmatch,
  )
 import Plutarch.Internal.Term (
+  S,
   Term,
   phoistAcyclic,
   plet,
@@ -53,11 +61,6 @@ infix 4 #==
 instance PEq PBool where
   {-# INLINEABLE (#==) #-}
   x #== y' = plet y' $ \y -> pif' # x # y #$ pnot # y
-
--- | @since WIP
-instance PEq PInteger where
-  {-# INLINEABLE (#==) #-}
-  x #== y = punsafeBuiltin PLC.EqualsInteger # x # y
 
 -- Helpers
 
@@ -93,3 +96,58 @@ pands ts' =
   case nonEmpty ts' of
     Nothing -> pcon PTrue
     Just ts -> foldl1 (#&&) ts
+
+-- | @since WIP
+instance PEq PInteger where
+  {-# INLINEABLE (#==) #-}
+  x #== y = punsafeBuiltin PLC.EqualsInteger # x # y
+
+instance PEq PData where
+  x #== y = punsafeBuiltin PLC.EqualsData # x # y
+
+instance PEq (PAsData a) where
+  x #== y = punsafeBuiltin PLC.EqualsData # x # y
+
+type family F (a :: S -> Type) :: Bool where
+  F PData = 'True
+  F (PAsData _) = 'True
+  F _ = 'False
+
+class Fc (x :: Bool) (a :: S -> Type) where
+  fc :: Proxy x -> Term s (PBuiltinList a) -> Term s (PBuiltinList a) -> Term s PBool
+
+instance (PEq a, PLC.Contains PLC.DefaultUni (PlutusRepr a)) => Fc 'False a where
+  fc _ xs ys = plistEquals # xs # ys
+    where
+      -- TODO: This is copied from ListLike. See if there's a way to not do this
+      plistEquals =
+        phoistAcyclic $
+          pfix #$ plam $ \self xlist ylist ->
+            pelimList
+              ( \x xs ->
+                  pelimList (\y ys -> pif (x #== y) (self # xs # ys) (pconstant False)) (pconstant False) ylist
+              )
+              (pelimList (\_ _ -> pconstant False) (pconstant True) ylist)
+              xlist
+
+instance PIsData (PBuiltinList a) => Fc 'True a where
+  fc _ xs ys = pdata xs #== pdata ys
+
+instance Fc (F a) a => PEq (PBuiltinList a) where
+  (#==) = fc (Proxy @(F a))
+
+instance (PEq a, PEq b) => PEq (PBuiltinPair a b) where
+  p1 #== p2 = pfstBuiltin # p1 #== pfstBuiltin # p2 #&& psndBuiltin # p1 #== psndBuiltin # p2
+
+instance PEq PByteString where
+  x #== y = punsafeBuiltin PLC.EqualsByteString # x # y
+
+-- | @since WIP
+instance PEq PByte where
+  {-# INLINEABLE (#==) #-}
+  x #== y = punsafeBuiltin PLC.EqualsInteger # x # y
+
+deriving anyclass instance PEq PLogicOpSemantics
+
+instance PEq PUnit where
+  x #== y = plet x \_ -> plet y \_ -> pcon PTrue
