+++ Plutarch/String.hs
@@ -17,13 +17,14 @@ import Data.Text (Text)
 import Data.Text qualified as Text
 import GHC.Generics (Generic)
 import Plutarch.Builtin.Bool (PBool, (#&&), (#||))
+import Plutarch.Builtin.ByteString (PByteString)
 import Plutarch.Builtin.Integer (PInteger)
-import Plutarch.ByteString (PByteString)
+import Plutarch.Builtin.Opaque (POpaque)
 import Plutarch.Internal.Eq (PEq ((#==)))
 import Plutarch.Internal.Lift (DeriveBuiltinPLiftable, PLiftable, PLifted (PLifted), pconstant)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)
+import Plutarch.Internal.Numeric
 import Plutarch.Internal.Ord ((#<=))
-import Plutarch.Internal.Other (POpaque)
 import Plutarch.Internal.PLam (plam)
 import Plutarch.Internal.PlutusType (DPTStrat, DerivePlutusType, PlutusType)
 import Plutarch.Internal.Term (
