+++ Plutarch/String.hs-boot
@@ -1,6 +1,6 @@
 module Plutarch.String (PString) where
 
-import Plutarch.Internal.Other (POpaque)
+import Plutarch.Builtin.Opaque (POpaque)
 import Plutarch.Internal.Term (Term)
 
 newtype PString s = PString (Term s POpaque)
